llm-evaluate/pyproject.toml
[project]
name = "llm-evaluate"
version = "0.1.0"
description = "Evaluation of LLM Models"
authors = [
    { name = "Brett Calcott", email = "brett@dragonfly.co.nz" },
]
dependencies = [
    "tortoise-orm>=0.21.3",
    "aiohttp>=3.9.5",
    "python-dotenv>=1.0.1",
    "stencila-types>=2.0.0a31",
    "typer>=0.12.3",
    "beautifulsoup4>=4.12.3",
    "pydantic>=2.7.3",
    "polars>=0.20.31",
    "numpy>=1.26.4",
]
requires-python = "==3.12.*"
readme = "README.md"

[project.license]
text = "Apache-2.0"

[project.scripts]
llm = "llm_evaluate:cli.main"

[build-system]
requires = [
    "pdm-backend",
]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest-asyncio>=0.23.7",
    "aioresponses>=0.7.6",
    "ruff>=0.4.8",
    "pytest>=8.2.2",
    "hypothesis>=6.103.1",
    "hypothesis-jsonschema>=0.23.1",
    "pytest-datadir>=1.5.0",
    "jupyterlab>=4.2.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyright]
include = [
    "src",
    "tests",
]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[tool.ruff]
target-version = "py312"
line-length = 88

src = [
    "src",
    "tests",
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "N",
    "UP",
    "S",
    "B",
    "A",
    "C4",
    "PIE",
    "ISC",
    "T20",
    "PT",
    "RET",
    "SIM",
    "PTH",
    "Q",
    "PLR",
    "PLE",
    "I001",
]
ignore = [
    "N812",
    "RET504",
    "ISC001",
]
unfixable = [
    "T20",
    "F841",
    "F401",
]
